% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvFunCrossMark.R
\name{EnvFunCrossMark}
\alias{EnvFunCrossMark}
\title{Function to obtain the Auto- and cross-function mark summary characteristics for the mark variogram and mark correlation envelopes}
\usage{
EnvFunCrossMark(
  xy,
  r,
  rf,
  f1,
  f2,
  stoyan = 0.15,
  TestFunc = "variogram",
  kernel = "epanechnikov",
  bw,
  lambda,
  correction = "isotropic",
  nsim = 199,
  nrank = 5,
  verbose = TRUE,
  Figure = TRUE,
  miny,
  maxy,
  Parallel = TRUE
)
}
\arguments{
\item{xy, }{object of class ppp (point pattern)}

\item{r, }{sequence of values to evaluate the  Auto- and cross-function mark summary characteristics}

\item{rf, }{sequence of values for the function associated to each point}

\item{f1}{is a matrix that contains de functions for each point, funct\link{,1} is the function for point 1 (first set of functions)}

\item{f2}{is a matrix that contains de functions for each point, funct\link{,1} is the function for point 1 (second set of functions)}

\item{stoyan}{value for the epanechnikov kernel, default stoyan=0.15}

\item{TestFunc, }{Test function. This can be TestFunc="variogram" or TestFunc="correlation"}

\item{kernel, }{kernel function for both the pair correlation function and the Functional mark correlation function}

\item{bw}{bandwidth for the kernel function for the pair correlation function and Functional mark correlation function}

\item{lambda, }{point intensity}

\item{correction}{for edge-effect for both the pair correlation function and the Functional mark correlation function}

\item{nsim, }{number of permutations for the Monte Carlo test, default nsim=299}

\item{nrank, Integer.}{Rank of the envelope value amongst the ‘nsim’ simulated values. A rank of 1 means that the minimum and maximum simulated values will be used}

\item{verbose, }{show the simulation progression, only for noParallel codification, default verbose=TRUE}

\item{Figure, }{display resulting FMCF and related max and min evelops. default Figure=TRUE}

\item{miny, }{maxy, plot parameters, ylim=c(miny,maxy)}

\item{Parallel, }{If 'TRUE' the algorithm will be considered assuming a parallel approach using the parallel package and assuming the total number of cores of your computer - 1.}

\item{alpha, }{signification level of the Monte Carlo Test}
}
\value{

}
\description{
This function computes an estimator of this function as described in Eckardt, Comas and Mateu (2023)
This function admits a parallel version, where the total number of cores -1 are used
}
\examples{
  To be implemented
 
}
\author{
Carles Comas \email{carles.comas@udl.cat}
}
