% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCrossMark.R
\name{FunCrossMark}
\alias{FunCrossMark}
\title{Function to obtain the Auto- and cross-function mark summary characteristics for the mark variogram and mark correlation}
\usage{
FunCrossMark(
  xy,
  r,
  rf,
  f1,
  f2,
  stoyan = 0.15,
  TestFunc = "variogram",
  kernel = "epanechnikov",
  bw,
  lambda,
  correction = "isotropic"
)
}
\arguments{
\item{xy, }{object of class ppp (point pattern)}

\item{r, }{sequence of values to evaluate the Auto- and cross-function mark summary characteristics}

\item{rf, }{sequence of values for the function associated to each point}

\item{f1}{is a matrix that contains the first set of functions for each point, funct1\link{,1} is the function of set 1 for point 1}

\item{f2}{is a matrix that contains the second set of functions for each point, funct1\link{,1} is the function of set 2 for point 1}

\item{stoyan}{value for the epanechnikov kernel, default stoyan=0.15}

\item{TestFunc, }{Test function. This can be TestFunc="variogram" or TestFunc="correlation"}

\item{kernel, }{kernel function for both the pair correlation function and the function mark summary characteristic}

\item{bw}{bandwidth for the kernel function for the pair correlation function and Functional mark correlation function}

\item{lambda, }{point intensity}

\item{correction}{for edge-effect for both the pair correlation function and the Functional mark correlation function}
}
\value{

}
\description{
This function computes an estimator of this function as described in Eckardt, Comas and Mateu (2023)
}
\examples{
 using Data_Duke
 Functiona.mark.cor(
 
}
\author{
Carles Comas \email{carles.comas@udl.cat}
}
